Porting sikuli environment from Chrome to Firefox
We used chromium-browser before. Chrome is fast, but unstable when running TW2 scripts.
I recognized that the script does not work sometimes in Chrome.
So I modified the sikuli script for Firefox.
I assume the user already stored his id/passwd in Firefox.
So we don't need to write password in the script from now.